Battleship
Turns are being announced, although I feel this is pointless based on the speed in which the computer moves. Winner is being announced in the DOM but the Reset Game is not resetting the game properly.

I need to look into this, then add functionality for people to place ships at specific locations (drag and drop)